need to fix price range

+++ contracts/interfaces/IRentPool.sol
@@ -7,7 +7,7 @@ interface IRentPool {
     function mint(address to) external returns (uint liquidity);
     function burn(address to) external returns (uint amountOfTokens, uint amountOfFees);
     function withdrawFees (address to) external returns (uint256 amountOfFees);
-    function getReserves() external view returns (uint112 _reserve, uint112 feesAccrued, uint32 _blockTimestampLast);
+    function getReserves() external view returns (uint112 _reserve, uint256 feesAccrued, uint32 _blockTimestampLast);
     function skim(address to) external;
     function sync() external;
